Good Progress
Finished up the module. The rest of it will be minor cosmetic tweaks, or new templates, with the addition of click tag support for flash.

Currently the module does not track click throughs. That's coming shortly.

cleaned up a bunch of other bits of code.

Advertisements and advertisement screens are a tad bit more human.